Fixes based on review: changing class object properties moved to public methods inside classes, added weak links in a closure, reorganized and marked all viewControllers for better readability, removed unnecessary comments, moved models to standalone files
