+++ ImageFeed/AuthView/AuthViewController.swift
@@ -12,9 +12,11 @@ protocol AuthViewControllerDelegate: AnyObject {
 }
 
 final class AuthViewController: UIViewController {
-    
+
+    // MARK: - Public Properties
     weak var delegate: AuthViewControllerDelegate?
-    
+
+    // MARK: - Private Properties
     private lazy var logoView: UIImageView = {
         let view = UIImageView()
         view.translatesAutoresizingMaskIntoConstraints = false
@@ -42,7 +44,8 @@ final class AuthViewController: UIViewController {
         
         return button
     }()
-    
+
+    // MARK: - Overrides Methods
     override func viewDidLoad() {
         super.viewDidLoad()
         
@@ -53,7 +56,8 @@ final class AuthViewController: UIViewController {
         setConstraints(for: logoView)
         setConstraints(for: signinButton)
     }
-    
+
+    // MARK: - Private Methods
     @objc private func signinButtonTapped() {
         navigateToWebView()
     }
@@ -84,6 +88,7 @@ final class AuthViewController: UIViewController {
     }
 }
 
+// MARK: - WebViewViewControllerDelegate
 extension AuthViewController: WebViewViewControllerDelegate {
     func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String) {
         vc.dismiss(animated: true)
